Captcha Breaking/dm_captcha/requirements.txt
numpy==1.16.0
opencv-python==4.0.0.21
Pillow==5.4.1
pytesseract==0.2.6
